improvements and game structure. still need moving BG

+++ Drawable.hpp
@@ -42,8 +42,8 @@ protected:
 	float xpos;
 	float ypos;
 
-	int xwid;
-	int ywid;
+	float xwid;
+	float ywid;
 	std::string sprite;
 	
 	char symbol;
